Move scope to base
