+++ luchador/nn/base/__init__.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import
 # pylint: disable=wildcard-import
 from .cost import *  # noqa: F401, F403
 from .layer import *  # noqa: F401, F403
+from .scope import *  # noqa: F401, F403
 from .wrapper import *  # noqa: F401, F403
 from .optimizer import *  # noqa: F401, F403
 from .initializer import *  # noqa: F401, F403
